Updated AWS and VMUG groups
This request was done on behalf of Brian Krausen. Updated AWS and VMUG User groups with bits of information. The VMUG calendar is located https://www.vmug.com/p/co/ly/gid=122/ but is located behind a login wall. Will defer to him on how he wants to proceed since it's his group

Signed-off-by: Kendrick Coleman <kendrickcoleman@gmail.com>